ChaosMod: Enable (almost) all warnings for GCC/Clang & treat all of them
Decrease the warning level of MSVC to W2 and mark all warnings as errors for all compilers.

Also use this ooportunity to finally raise the cmake version.

+++ ChaosMod/Effects/db/Peds/PedsKillerClowns.cpp
@@ -9,7 +9,7 @@
 CHAOS_VAR std::list<Ped> clownEnemies;
 CHAOS_VAR int spawnTimer = -1;
 CHAOS_VAR Hash relationshipGroup;
-CHAOS_VAR int maxClownsToSpawn = 3;
+CHAOS_VAR size_t maxClownsToSpawn = 3;
 
 static Vector3 getRandomOffsetCoord(Vector3 startCoord, int minOffset, int maxOffset)
 {
