+++ ChaosMod/Effects/db/Peds/PedsSmokeTrails.cpp
@@ -20,7 +20,7 @@ static void OnStart()
 
 static void OnStop()
 {
-	for (int i = 0; i < peds.size(); i++)
+	for (size_t i = 0; i < peds.size(); i++)
 	{
 		REMOVE_PARTICLE_FX_FROM_ENTITY(peds.at(i));
 		peds.erase(peds.begin() + i);
@@ -36,7 +36,8 @@ static void OnTick()
 		{
 			USE_PARTICLE_FX_ASSET(ptfxDict);
 
-			int ptfx = START_PARTICLE_FX_LOOPED_ON_ENTITY(ptfxName, ped, 0, -0.1, 0, 0, 0, 0, 0.5, false, false, false);
+			int ptfx =
+			    START_PARTICLE_FX_LOOPED_ON_ENTITY(ptfxName, ped, 0, -0.1f, 0, 0, 0, 0, 0.5f, false, false, false);
 
 			SET_PARTICLE_FX_LOOPED_COLOUR(ptfx, g_Random.GetRandomFloat(0, 1), g_Random.GetRandomFloat(0, 1),
 			                              g_Random.GetRandomFloat(0, 1), false);
